adding load balance system

+++ src/particleGen/specific/includes/gammaPolarised.hh
@@ -1,8 +1,8 @@
 
 //
 //
-//    Copyright (C) 2019 Universitat de València - UV
-//    Copyright (C) 2019 Universitat Politècnica de València - UPV
+//    Copyright (C) 2019-2020 Universitat de València - UV
+//    Copyright (C) 2019-2020 Universitat Politècnica de València - UPV
 //
 //    This file is part of PenRed: Parallel Engine for Radiation Energy Deposition.
 //
@@ -49,7 +49,7 @@ class gammaPolarised_specificSampler : public abc_specificSampler<pen_state_gPol
   
   void sample(pen_state_gPol& state,
 	      pen_KPAR& /*genKpar*/,
-	      unsigned long& /*dhist*/,
+	      unsigned long long& /*dhist*/,
 	      pen_rand& /*random*/);
   
   int configure(double& /*Emax*/,
